+++ tests/testthat/test-sort_hierarchical.R
@@ -7,13 +7,13 @@ tbl <- tbl_hierarchical(
   data = ADAE_subset,
   variables = c(SEX, RACE, AETERM),
   by = TRTA,
-  denominator = cards::ADSL |> mutate(TRTA = ARM),
+  denominator = cards::ADSL,
   id = USUBJID,
   overall_row = TRUE
 )
 
 test_that("sort_hierarchical() works", {
-  withr::local_options(width = 200)
+  withr::local_options(width = 250)
 
   # no errors
   expect_silent(tbl <- sort_hierarchical(tbl))
@@ -68,12 +68,46 @@ test_that("sort_hierarchical(sort = 'alphanumeric') works", {
   )
 })
 
+test_that("sort_hierarchical(sort) works with different sorting methods for each variable", {
+  expect_silent(
+    tbl <- sort_hierarchical(
+      tbl,
+      sort = list(SEX ~ "alphanumeric", RACE = "descending", AETERM = "alphanumeric")
+    )
+  )
+
+  # results are ordered correctly
+  expect_equal(
+    tbl$table_body |>
+      dplyr::filter(variable == "SEX") |>
+      dplyr::pull(label),
+    c("F", "M")
+  )
+  expect_equal(
+    tbl$table_body |>
+      dplyr::filter(variable == "RACE") |>
+      dplyr::pull(label),
+    c("WHITE", "BLACK OR AFRICAN AMERICAN", "WHITE", "BLACK OR AFRICAN AMERICAN", "AMERICAN INDIAN OR ALASKA NATIVE")
+  )
+  expect_equal(
+    tbl$table_body |>
+      dplyr::filter(variable == "AETERM") |>
+      dplyr::pull(label),
+    c(
+      "APPLICATION SITE ERYTHEMA", "APPLICATION SITE PRURITUS", "DIARRHOEA", "ERYTHEMA", "APPLICATION SITE PRURITUS",
+      "ATRIOVENTRICULAR BLOCK SECOND DEGREE", "DIARRHOEA", "ERYTHEMA", "APPLICATION SITE ERYTHEMA",
+      "APPLICATION SITE PRURITUS", "ATRIOVENTRICULAR BLOCK SECOND DEGREE", "DIARRHOEA", "ERYTHEMA",
+      "APPLICATION SITE PRURITUS", "DIARRHOEA", "ERYTHEMA", "ERYTHEMA"
+    )
+  )
+})
+
 test_that("sort_hierarchical() works when there is no overall row in x", {
   tbl_no_overall <- tbl_hierarchical(
     data = ADAE_subset,
     variables = c(SEX, RACE, AETERM),
     by = TRTA,
-    denominator = cards::ADSL |> mutate(TRTA = ARM),
+    denominator = cards::ADSL,
     id = USUBJID,
     overall_row = FALSE
   )
@@ -98,7 +132,7 @@ test_that("sort_hierarchical() works with only one variable in x", {
     data = ADAE_subset,
     variables = AETERM,
     by = TRTA,
-    denominator = cards::ADSL |> mutate(TRTA = ARM),
+    denominator = cards::ADSL,
     id = USUBJID,
     overall_row = TRUE
   )
@@ -130,7 +164,7 @@ test_that("sort_hierarchical() works when some variables not included in x", {
     data = ADAE_subset,
     variables = c(SEX, RACE, AETERM),
     by = TRTA,
-    denominator = cards::ADSL |> mutate(TRTA = ARM),
+    denominator = cards::ADSL,
     id = USUBJID,
     include = c(SEX, AETERM),
     overall_row = TRUE
@@ -189,6 +223,22 @@ test_that("sort_hierarchical() works with add_overall()", {
       "APPLICATION SITE PRURITUS", "DIARRHOEA", "ERYTHEMA", "ERYTHEMA"
     )
   )
+
+  # overall col with different sort variables does not affect sort order
+  expect_silent(
+    tbl_s <- sort_hierarchical(
+      tbl,
+      sort = list(SEX ~ "alphanumeric", RACE = "descending", AETERM = "alphanumeric")
+    )
+  )
+  tbl_o <- tbl |> add_overall()
+  expect_silent(
+    tbl_o <- sort_hierarchical(
+      tbl_o,
+      sort = list(SEX ~ "alphanumeric", RACE = "descending", AETERM = "alphanumeric")
+    )
+  )
+  expect_identical(tbl_o$table_body$label, tbl_s$table_body$label)
 })
 
 test_that("sort_hierarchical() error messaging works", {