+++ tests/testthat/test-tbl_ard_hierarchical.R
@@ -12,7 +12,7 @@ test_that("tbl_ard_hierarchical() event rates", {
       data = ADAE_subset,
       variables = c(AESOC, AETERM, AESEV),
       by = TRTA,
-      denominator = cards::ADSL |> mutate(TRTA = ARM),
+      denominator = cards::ADSL,
       id = USUBJID
     )
 
@@ -38,7 +38,7 @@ test_that("tbl_ard_hierarchical() event rates", {
       data = ADAE_subset,
       variables = c(AESOC, AETERM, AESEV),
       by = TRTA,
-      denominator = cards::ADSL |> mutate(TRTA = ARM),
+      denominator = cards::ADSL,
       id = USUBJID,
       digits = ~list(p = 1L)
     ) |>
@@ -52,7 +52,7 @@ test_that("tbl_ard_hierarchical() counts", {
       data = ADAE_subset,
       variables = c(AESOC, AETERM),
       by = TRTA,
-      denominator = cards::ADSL |> mutate(TRTA = ARM)
+      denominator = cards::ADSL
     )
 
   expect_error(
@@ -71,8 +71,40 @@ test_that("tbl_ard_hierarchical() counts", {
       data = ADAE_subset,
       variables = c(AESOC, AETERM),
       by = TRTA,
-      denominator = cards::ADSL |> mutate(TRTA = ARM)
+      denominator = cards::ADSL
     ) |>
       as.data.frame(col_labels = FALSE)
+  )
+})
+
+test_that("tbl_ard_hierarchical() preserves ARD sorting", {
+  ard <-
+    cards::ard_stack_hierarchical(
+      data = ADAE_subset,
+      variables = c(AESOC, AETERM),
+      by = TRTA,
+      id = USUBJID,
+      denominator = cards::ADSL
+    ) |>
+    cards::sort_ard_hierarchical("descending")
+
+  expect_silent(
+    tbl <- tbl_ard_hierarchical(
+      cards = ard,
+      variables = c(AESOC, AETERM),
+      by = TRTA
+    )
+  )
+
+  expect_equal(
+    tbl$table_body |>
+      select("variable", "label"),
+    ard |>
+      dplyr::filter(variable != "TRTA") |>
+      cards::unlist_ard_columns() |>
+      select(cards::all_ard_variables()) |>
+      dplyr::rename(label = variable_level) |>
+      dplyr::distinct(),
+    ignore_attr = TRUE
   )
 })